requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
python-dotenv==1.0.1
streamlit==1.38.0
pydantic==2.9.0
dashscope
chromadb==0.5.5
pandas==2.2.2 